Move events from service/ to worker/
They're not used anywhere in the service/.
