+++ packit_service/worker/events/koji.py
@@ -14,7 +14,7 @@ from packit_service.models import (
     ProjectReleaseModel,
     GitBranchModel,
 )
-from packit_service.service.events.event import AbstractForgeIndependentEvent
+from packit_service.worker.events.event import AbstractForgeIndependentEvent
 
 
 class KojiBuildEvent(AbstractForgeIndependentEvent):
