+++ packit_service/worker/testing_farm.py
@@ -17,7 +17,7 @@ from packit_service.config import ServiceConfig
 from packit_service.constants import TESTING_FARM_INSTALLABILITY_TEST_URL
 from packit_service.models import CoprBuildModel, TFTTestRunModel, TestingFarmResult
 from packit_service.sentry_integration import send_to_sentry
-from packit_service.service.events import EventData
+from packit_service.worker.events import EventData
 from packit_service.service.urls import get_testing_farm_info_url
 from packit_service.worker.build import CoprBuildJobHelper
 from packit_service.worker.result import TaskResults
