+++ tests/conftest.py
@@ -12,7 +12,7 @@ from ogr import GithubService, GitlabService
 from packit.config import JobConfigTriggerType
 from packit_service.config import ServiceConfig
 from packit_service.models import JobTriggerModelType
-from packit_service.service.events import (
+from packit_service.worker.events import (
     PullRequestGithubEvent,
     PushGitHubEvent,
     ReleaseEvent,
