+++ tests_requre/database/test_tasks.py
@@ -16,7 +16,7 @@ from packit_service.models import (
     SRPMBuildModel,
     PullRequestModel,
 )
-from packit_service.service.events import AbstractCoprBuildEvent
+from packit_service.worker.events import AbstractCoprBuildEvent
 from packit_service.worker.build.babysit import check_copr_build
 
 BUILD_ID = 1300329
